data_connections/requirements.txt
aiofiles==24.1.0
alembic==1.16.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroalign==2.6.2
astrometry==4.3.0
astropy==7.0.1
astropy-iers-data==0.2025.5.12.0.38.29
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bokeh==2.4.3
boto3==1.38.13
botocore==1.38.13
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.2
cycler==0.12.1
dacite==1.7.0
databricks-sdk==0.55.0
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
docker==7.1.0
docutils==0.21.2
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastapi==0.115.12
fastjsonschema==2.21.1
fiftyone==1.5.2
fiftyone-brain==0.21.2
fiftyone_db==1.1.7
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
fqdn==1.5.1
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
glob2==0.7
google-auth==2.40.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
inflate64==1.0.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipympl==0.9.8
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1745672166/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.8
lazy_loader==0.4
loro==1.5.1
Mako==1.3.10
marimo==0.13.15
Markdown==3.8
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mistune==3.1.3
mlflow==2.22.0
mlflow-skinny==2.22.0
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
notebook==7.4.2
notebook_shim==0.2.4
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
photutils==2.2.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
plotly==6.1.0
pprintpp==0.4.0
priority==2.0.0
prometheus_client==0.21.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
protobuf==3.20.3
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py7zr==0.22.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycocotools==2.0.10
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydash==8.0.5
pydeck==0.9.1
pyerfa==2.0.1.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pymdown-extensions==10.15
pymongo==4.9.2
Pympler==1.1
pyparsing==3.2.3
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1734687138743/work
pyzstd==0.17.0
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.25.0
rsa==4.9.1
rtree==1.4.0
s3transfer==0.12.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
Send2Trash==1.8.3
sep==1.4.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==0.10.3
sseclient-py==1.8.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.2
strawberry-graphql==0.270.0
streamlit==1.22.0
streamlit-bokeh==3.8.0
streamlit-drawable-canvas==0.9.3
streamlit-img-label==0.1.1
streamlit-labelstudio==0.0.3
streamlit_image_annotation==0.7.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
toolz==1.1.0
torch==2.7.1
torchvision==0.22.1
tornado @ file:///croot/tornado_1733960490606/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
triton==3.3.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.3
validators==0.35.0
voxel51-eta==0.14.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
xyzservices==2025.4.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
